Update schedule.rb
Hey, i'm using this great tool with a lot of scheduled jobs in a sqlite3 database; So i ran into trouble because everytime api/schedule is called, the app was blocked, because of the locked database while updating ALL jobs.
So, i decided to limit the number of updated jobs and change the order by state, too. 
Maybe its more intelligent to change the limit of updated jobs to the number of hosts/slots available.